Option to use or not the first VBox interface. Object rotation with keypad and '+' / '-'  (testing needed). Debug level is saved. Kernel options support double quotes and are protected with double quotes when passing the argument to Qemu.
Global TCP port tracker to allocate ports and avoid conflicts in GNS3.
	- Check if ports are already in use by other programs
	- Console + AUX ports for IOS routers are tracked
	- Console ports for Qemu-like devices are tracked
	- Command 'show ports' to display all port allocations.
	- Level one debug have been added.

Tracker algorithm is to find a port not used in range port to port+5, then try random port in range port to port+100 (10 tries by default, each time trying to find a port not used in range random_port+5). All of this while never allocating a port already used by GNS3. This give total tries to 10*5 = 50. Finally if no port can be found, return the original port and hope for the best (old behaviour).

+++ src/GNS3/Node/ATMBR.py
@@ -71,8 +71,9 @@ class ATMBR(AbstractNode):
 
         if self.atmbr:
             try:
-                self.atmbr.delete()
-                del self.dynagen.devices[self.hostname]
+                if self.dynagen.devices.has_key(self.hostname):
+                    self.atmbr.delete()
+                    del self.dynagen.devices[self.hostname]
                 if self.atmbr in self.hypervisor.devices:
                     self.hypervisor.devices.remove(self.atmbr)
                 self.dynagen.update_running_config()
