+++ src/GNS3/Node/IOSRouter.py
@@ -135,8 +135,9 @@ class IOSRouter(AbstractNode):
             try:
                 self.stopNode()
                 # don't forget to delete this router in Dynamips
-                self.router.delete()
-                del self.dynagen.devices[self.hostname]
+                if self.dynagen.devices.has_key(self.hostname):
+                    self.router.delete()
+                    del self.dynagen.devices[self.hostname]
                 if self.router in self.hypervisor.devices:
                     self.hypervisor.devices.remove(self.router)
                 self.dynagen.update_running_config()
